Move export related code to file, rename test
